refactor: share sentry mocks and types
