+++ apps/frontend/src/types/sentry-electron.d.ts
@@ -1,49 +1,32 @@
+interface SentryErrorEvent {
+  [key: string]: unknown;
+}
+
+interface SentryScope {
+  setContext: (key: string, value: Record<string, unknown>) => void;
+}
+
+interface SentryInitOptions {
+  beforeSend?: (event: SentryErrorEvent) => SentryErrorEvent | null;
+  tracesSampleRate?: number;
+  profilesSampleRate?: number;
+  dsn?: string;
+  environment?: string;
+  release?: string;
+  debug?: boolean;
+  enabled?: boolean;
+}
+
 declare module '@sentry/electron/main' {
-  export interface ErrorEvent {
-    [key: string]: unknown;
-  }
-
-  export interface Scope {
-    setContext: (key: string, value: Record<string, unknown>) => void;
-  }
-
-  export interface InitOptions {
-    beforeSend?: (event: ErrorEvent) => ErrorEvent | null;
-    tracesSampleRate?: number;
-    profilesSampleRate?: number;
-    dsn?: string;
-    environment?: string;
-    release?: string;
-    debug?: boolean;
-    enabled?: boolean;
-  }
-
-  export function init(options: InitOptions): void;
+  export type ErrorEvent = SentryErrorEvent;
+  export function init(options: SentryInitOptions): void;
   export function captureException(error: Error): void;
-  export function withScope(callback: (scope: Scope) => void): void;
+  export function withScope(callback: (scope: SentryScope) => void): void;
 }
 
 declare module '@sentry/electron/renderer' {
-  export interface ErrorEvent {
-    [key: string]: unknown;
-  }
-
-  export interface Scope {
-    setContext: (key: string, value: Record<string, unknown>) => void;
-  }
-
-  export interface InitOptions {
-    beforeSend?: (event: ErrorEvent) => ErrorEvent | null;
-    tracesSampleRate?: number;
-    profilesSampleRate?: number;
-    dsn?: string;
-    environment?: string;
-    release?: string;
-    debug?: boolean;
-    enabled?: boolean;
-  }
-
-  export function init(options: InitOptions): void;
+  export type ErrorEvent = SentryErrorEvent;
+  export function init(options: SentryInitOptions): void;
   export function captureException(error: Error): void;
-  export function withScope(callback: (scope: Scope) => void): void;
+  export function withScope(callback: (scope: SentryScope) => void): void;
 }
